\name{tolerance-package}
\alias{tolerance-package}
\alias{tolerance}
\docType{package}
\title{
Functions for Calculating Tolerance Intervals
}
\description{
A collection of functions for calculating \code{[100(1-alpha)\%, 100(P)\%]} tolerance intervals, which
are intervals with \code{100(1-alpha)\%} confidence of covering \code{100(P)\%} of the population of interest.
}
\details{
\tabular{ll}{
Package: \tab tolerance\cr
Type: \tab Package\cr
Version: \tab 0.5.1\cr
Date: \tab 2012-08-28\cr
Depends \tab rgl, stats4\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Derek S. Young, Ph.D.

Maintainer: Derek S. Young <dsy109@stat.psu.edu>
}
\seealso{
  \code{\link{confint}}
}
\references{
  Hahn, G. J. and Meeker, W. Q. (1991), \emph{Statistical Intervals: A Guide for Practitioners}, Wiley-Interscience.

  Krishnamoorthy, K. and Mathew, T. (2009), \emph{Statistical Tolerance Regions: Theory, Applications, and Computation}, Wiley.

  Patel, J. K. (1986), Tolerance Intervals - A Review, \emph{Communications in Statistics - Theory and Methodology},
  \bold{15}, 2719--2762.
  
  Young, D. S. (2010), \code{tolerance:} An \code{R} Package for Estimating Tolerance Intervals, \emph{Journal of Statistical Software}, \bold{36}(5), 1--39.
}

