\name{tolerance-package}
\alias{tolerance-package}
\alias{tolerance}
\docType{package}
\title{
Functions for Calculating Tolerance Intervals
}
\description{
A collection of functions for calculating \code{[100(1-alpha)\%, 100(P)\%]} tolerance intervals, which
are intervals with \code{100(1-alpha)\%} confidence of covering \code{100(P)\%} of the population of interest.
}
\details{
\tabular{ll}{
Package: \tab tolerance\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2009-04-07\cr
Depends \tab rgl\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Derek S. Young, Ph.D.

Maintainer: Derek S. Young <dsy109@stat.psu.edu>
}
\references{
  Hahn, G. J. and Meeker, W. Q. (1991), \emph{Statistical Intervals: A Guide for Practitioners}, Wiley-Interscience.

  Patel, J. K. (1986), Tolerance Intervals - A Review, \emph{Communications in Statistics - Theory and Methodology},
  \bold{15}, 2719--2762.
}

